Move helper function to own header
